refactored the implementation as the previous implementation was violating open-close and single responsibility principle. Improved code readablity and test coverage added, adding any new price method will be possible without affecting the current implementation
